Add documentation.
This should fix all but two validations errors that @charlesb flagged, namely:
- missing meta files
- missing CHANGELOG

It's also a start to the missing doc; Ron should take it from here.

The error about the runtime being available in the editor is not fixed, because it needs to be present in the editor. There's something we don't understand.